Reflekt models refactoring
