+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/analysis/ReflektModuleAnalysisExtension.kt
@@ -5,14 +5,12 @@ import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.com.intellij.openapi.project.Project
 import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.descriptors.impl.ModuleDescriptorImpl
-import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.resolve.BindingTrace
 import org.jetbrains.kotlin.resolve.jvm.extensions.AnalysisHandlerExtension
-import org.jetbrains.reflekt.plugin.analysis.analyzer.descriptor.DescriptorAnalyzer
 import org.jetbrains.reflekt.plugin.analysis.analyzer.source.ReflektAnalyzer
 import org.jetbrains.reflekt.plugin.analysis.models.*
-import org.jetbrains.reflekt.plugin.analysis.resolve.getDescriptors
+import org.jetbrains.reflekt.plugin.analysis.models.ir.*
 import org.jetbrains.reflekt.plugin.analysis.serialization.SerializationUtils
 import org.jetbrains.reflekt.plugin.generation.code.generator.ReflektImplGenerator
 import org.jetbrains.reflekt.plugin.utils.Util.getInstances
@@ -21,40 +19,40 @@ import org.jetbrains.reflekt.plugin.utils.Util.saveUses
 import java.io.File
 
 class ReflektModuleAnalysisExtension(
-    private val reflektMetaFiles: Set<File>,
+    private val reflektMetaFilesFromLibraries: Set<File>,
     private val toSaveMetadata: Boolean,
     private val generationPath: File?,
     private val reflektMetaFile: File?,
     private val reflektContext: ReflektContext? = null,
     private val messageCollector: MessageCollector? = null
 ) : AnalysisHandlerExtension {
+    private val reflektPackage = "org.jetbrains.reflekt"
 
     // TODO: store ReflektMetaInf by modules
     override fun analysisCompleted(project: Project, module: ModuleDescriptor, bindingTrace: BindingTrace, files: Collection<KtFile>): AnalysisResult? {
         messageCollector?.log("ReflektAnalysisExtension is starting...")
         (module as? ModuleDescriptorImpl) ?: error("Internal error! Can not cast a ModuleDescriptor to ModuleDescriptorImpl")
-        val (libraryInvokes, packages) = getReflektMeta(reflektMetaFiles, module)
+        val externalLibrariesAnalyzer = ExternalLibrariesAnalyzer(reflektMetaFilesFromLibraries, module, messageCollector)
 
-        // TODO: use a const
-        val setOfFiles = files.filter { it.packageFqName.asString() != "org.jetbrains.reflekt" }.toSet()
+        val setOfFiles = files.filter { it.packageFqName.asString() != reflektPackage }.toSet()
         val analyzer = ReflektAnalyzer(setOfFiles, bindingTrace.bindingContext, messageCollector)
         val invokes = analyzer.invokes()
         messageCollector?.log("Project's invokes: $invokes")
         if (toSaveMetadata && reflektMetaFile != null) {
             saveMetaData(invokes, setOfFiles)
         }
-        val mergedInvokes = invokes.merge(libraryInvokes)
+        val mergedInvokes = invokes.merge(externalLibrariesAnalyzer.invokesWithPackages.invokes)
         messageCollector?.log("Merged invokes: $mergedInvokes")
         val uses = analyzer.uses(mergedInvokes)
         bindingTrace.saveUses(uses)
 
         if (reflektContext != null) {
-            messageCollector?.log("Start analysis ${module.name} module's files")
-            val sourceUses = IrReflektUses.fromReflektUses(uses, bindingTrace.bindingContext)
-            val librariesUses = getUsesFromLibraries(module, packages, invokes)
-            reflektContext.uses = sourceUses.merge(librariesUses)
-            messageCollector?.log("IrReflektUses were created successfully")
+            reflektContext.uses = externalLibrariesAnalyzer.buildIrReflektUses(
+                uses,
+                bindingTrace.bindingContext,
+            )
 
+            // TODO: collect instances only if SmartReflekt calls exist
             // Need only for SmartReflekt
             val instances = getInstances(setOfFiles, bindingTrace, messageCollector = messageCollector)
             reflektContext.instances = IrReflektInstances.fromReflektInstances(instances, bindingTrace.bindingContext, messageCollector)
@@ -65,7 +63,7 @@ class ReflektModuleAnalysisExtension(
         }
 
         val reflektImplFile = File(generationPath, "io/reflekt/ReflektImpl.kt")
-        if (toGenerateReflektImpl(libraryInvokes)) {
+        if (toGenerateReflektImpl(externalLibrariesAnalyzer.invokesWithPackages.invokes)) {
             generateReflektImpl(uses, reflektImplFile)
         }
         messageCollector?.log("Finish analysis with ReflektModuleAnalysisExtension")
@@ -87,36 +85,6 @@ class ReflektModuleAnalysisExtension(
             )
         }
         messageCollector?.log("Finish generation ReflektImpl")
-    }
-
-    private fun getUsesFromLibraries(module: ModuleDescriptorImpl, packages: Set<String>, invokes: ReflektInvokes): IrReflektUses {
-        var uses = IrReflektUses()
-        module.getDescriptors(packages.map { FqName(it) }.toSet()).forEach {
-            val ms = it.getMemberScope()
-            val currentUses = DescriptorAnalyzer(ms, messageCollector).uses(invokes)
-            messageCollector?.log("CURRENT LIBRARY USES: $currentUses")
-            uses = uses.merge(currentUses)
-        }
-        return uses
-    }
-
-    private fun getReflektMeta(reflektMetaFiles: Set<File>, module: ModuleDescriptorImpl): ReflektInvokesWithPackages {
-        messageCollector?.log("reflektMetaFiles $reflektMetaFiles")
-        var libraryInvokes = ReflektInvokes()
-        val packages = mutableSetOf<String>()
-        reflektMetaFiles.forEach {
-            val currentInvokesWithPackages = SerializationUtils.decodeInvokes(it.readBytes(), module)
-            messageCollector?.log("Deserialized invokes: ${currentInvokesWithPackages.invokes}")
-            messageCollector?.log("Deserialized packages: ${currentInvokesWithPackages.packages}")
-            libraryInvokes = libraryInvokes.merge(currentInvokesWithPackages.invokes)
-            packages.addAll(currentInvokesWithPackages.packages)
-        }
-        messageCollector?.log("Library invokes: $libraryInvokes")
-        messageCollector?.log("Library packages: $packages")
-        return ReflektInvokesWithPackages(
-            invokes = libraryInvokes,
-            packages = packages
-        )
     }
 
     private fun saveMetaData(invokes: ReflektInvokes, files: Set<KtFile>) {
