+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/analysis/models/ir/IrCommon.kt
@@ -1,0 +1,6 @@
+package org.jetbrains.reflekt.plugin.analysis.models.ir
+
+data class ReflektContext(
+    var uses: IrReflektUses? = null,
+    var instances: IrReflektInstances? = null
+)
