+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/analysis/processor/source/instances/ClassOrObjectInstancesProcessor.kt
@@ -2,12 +2,13 @@ package org.jetbrains.reflekt.plugin.analysis.processor.source.instances
 
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.resolve.BindingContext
+import org.jetbrains.reflekt.plugin.analysis.models.TypeInstances
 import org.jetbrains.reflekt.plugin.analysis.processor.*
 
-open class ClassOrObjectInstancesProcessor<T : KtClassOrObject>(override val binding: BindingContext) : BaseInstancesProcessor<MutableList<T>>(binding) {
-    override val fileToInstances: HashMap<FileID, MutableList<T>> = HashMap()
+open class ClassOrObjectInstancesProcessor<T : KtClassOrObject>(override val binding: BindingContext) : BaseInstancesProcessor<TypeInstances<T>>(binding) {
+    override val fileToInstances: HashMap<FileID, TypeInstances<T>> = HashMap()
 
-    override fun process(element: KtElement, file: KtFile): HashMap<FileID, MutableList<T>> {
+    override fun process(element: KtElement, file: KtFile): HashMap<FileID, TypeInstances<T>> {
         (element as? T)?.let {
             fileToInstances.getOrPut(file.fullName) { ArrayList() }.add(it)
         }
