+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/analysis/processor/source/instances/FunctionInstancesProcessor.kt
@@ -2,12 +2,13 @@ package org.jetbrains.reflekt.plugin.analysis.processor.source.instances
 
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.resolve.BindingContext
+import org.jetbrains.reflekt.plugin.analysis.models.FunctionInstances
 import org.jetbrains.reflekt.plugin.analysis.processor.*
 
-class FunctionInstancesProcessor(override val binding: BindingContext) : BaseInstancesProcessor<MutableList<KtNamedFunction>>(binding) {
-    override val fileToInstances: HashMap<FileID, MutableList<KtNamedFunction>> = HashMap()
+class FunctionInstancesProcessor(override val binding: BindingContext) : BaseInstancesProcessor<FunctionInstances>(binding) {
+    override val fileToInstances: HashMap<FileID, FunctionInstances> = HashMap()
 
-    override fun process(element: KtElement, file: KtFile): HashMap<FileID, MutableList<KtNamedFunction>> {
+    override fun process(element: KtElement, file: KtFile): HashMap<FileID, FunctionInstances> {
         (element as? KtNamedFunction)?.let {
             fileToInstances.getOrPut(file.fullName) { ArrayList() }.add(it)
         }
