+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/analysis/processor/source/invokes/base/BaseClassInvokesProcessor.kt
@@ -1,15 +1,14 @@
-package org.jetbrains.reflekt.plugin.analysis.processor.source.invokes
+package org.jetbrains.reflekt.plugin.analysis.processor.source.invokes.base
 
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
-import org.jetbrains.kotlin.psi.*
+import org.jetbrains.kotlin.psi.KtElement
+import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.resolve.BindingContext
-import org.jetbrains.reflekt.plugin.analysis.common.ReflektEntity
 import org.jetbrains.reflekt.plugin.analysis.models.ClassOrObjectInvokes
 import org.jetbrains.reflekt.plugin.analysis.processor.FileID
 import org.jetbrains.reflekt.plugin.analysis.processor.fullName
-import org.jetbrains.reflekt.plugin.analysis.psi.getFqName
 
-class ClassInvokesProcessor(override val binding: BindingContext, override val messageCollector: MessageCollector?) :
+abstract class BaseClassInvokesProcessor(override val binding: BindingContext, override val messageCollector: MessageCollector?) :
     BaseInvokesProcessor<ClassOrObjectInvokes>(binding, messageCollector) {
     override val fileToInvokes: HashMap<FileID, ClassOrObjectInvokes> = HashMap()
 
@@ -20,5 +19,4 @@ class ClassInvokesProcessor(override val binding: BindingContext, override val m
         return fileToInvokes
     }
 
-    override fun isValidExpression(expression: KtReferenceExpression) = expression.getFqName(binding) == ReflektEntity.CLASSES.fqName
 }
