+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/analysis/processor/source/invokes/base/BaseFunctionInvokesProcessor.kt
@@ -1,17 +1,15 @@
-package org.jetbrains.reflekt.plugin.analysis.processor.source.invokes
+package org.jetbrains.reflekt.plugin.analysis.processor.source.invokes.base
 
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.resolve.BindingContext
 import org.jetbrains.kotlin.utils.addIfNotNull
-import org.jetbrains.reflekt.plugin.analysis.common.ReflektEntity
 import org.jetbrains.reflekt.plugin.analysis.common.findReflektFunctionInvokeArgumentsByExpressionPart
 import org.jetbrains.reflekt.plugin.analysis.models.FunctionInvokes
 import org.jetbrains.reflekt.plugin.analysis.processor.FileID
 import org.jetbrains.reflekt.plugin.analysis.processor.fullName
-import org.jetbrains.reflekt.plugin.analysis.psi.getFqName
 
-class FunctionInvokesProcessor(override val binding: BindingContext, override val messageCollector: MessageCollector?) :
+abstract class BaseFunctionInvokesProcessor(override val binding: BindingContext, override val messageCollector: MessageCollector?) :
     BaseInvokesProcessor<FunctionInvokes>(binding, messageCollector) {
     override val fileToInvokes: HashMap<FileID, FunctionInvokes> = HashMap()
 
@@ -21,6 +19,4 @@ class FunctionInvokesProcessor(override val binding: BindingContext, override va
         }
         return fileToInvokes
     }
-
-    override fun isValidExpression(expression: KtReferenceExpression) = expression.getFqName(binding) == ReflektEntity.FUNCTIONS.fqName
 }
