+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/analysis/resolve/DescriptorsUtil.kt
@@ -6,7 +6,7 @@ import org.jetbrains.kotlin.incremental.components.NoLookupLocation
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.resolve.descriptorUtil.*
 import org.jetbrains.kotlin.resolve.scopes.MemberScope
-import org.jetbrains.reflekt.plugin.analysis.models.IrFunctionInfo
+import org.jetbrains.reflekt.plugin.analysis.models.ir.IrFunctionInfo
 import org.jetbrains.reflekt.plugin.analysis.psi.function.*
 
 // Descriptors for classes and objects
