+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/generation/code/generator/ReflektImplGenerator.kt
@@ -2,7 +2,7 @@ package org.jetbrains.reflekt.plugin.generation.code.generator
 
 import com.squareup.kotlinpoet.ClassName
 import org.jetbrains.reflekt.plugin.analysis.models.ReflektUses
-import org.jetbrains.reflekt.plugin.analysis.models.flatten
+import org.jetbrains.reflekt.plugin.analysis.models.ir.flatten
 import org.jetbrains.reflekt.plugin.generation.code.generator.models.*
 import java.util.*
