+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/generation/ir/SmartReflektIrGenerationExtension.kt
@@ -4,7 +4,7 @@ import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
-import org.jetbrains.reflekt.plugin.analysis.models.ReflektContext
+import org.jetbrains.reflekt.plugin.analysis.models.ir.ReflektContext
 import java.io.File
 
 class SmartReflektIrGenerationExtension(
